mdbook-comments: Fix reply submission API endpoint inconsistency
The saveReply() function was using /api/comments endpoints while
loadComments() and saveComment() used /comments endpoints. This
caused replies to be saved to the wrong endpoint and not appear
in the UI after submission.

Fixed by making all endpoints consistent:
- loadComments(): /comments
- saveComment(): /comments  
- saveReply(): /comments (was /api/comments)

This fixes the 'should reply to a comment' test failure.

Fixes issues-7, issues-8, bd-356